Progress...

+++ src/test/java/cointoss/util/arithmetic/BigDecimalBenchmark.java
@@ -28,7 +28,7 @@ public class BigDecimalBenchmark {
 
         benchmark.measure("Num#calculate", () -> {
             return cointoss.util.arithmetic.Num.of(1000).calculate(Num.of(4), Num.of(0.123), Num.of(-24.5), (p0, p1, p2, p3) -> {
-                long v = ((p0 / p1) + p2) * p3;
+                double v = ((p0 / p1) + p2) * p3;
                 return v * v;
             });
         });
